backend/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
chromadb>=0.4.0

# Document parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Embeddings and LLM
openai>=1.12.0
sentence-transformers>=2.2.0

# ServiceNow connector
pysnow>=0.7.0

# SharePoint connector
Office365-REST-Python-Client>=2.5.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
